moved mockserver to src for sharing it with app module
