+++ core/build.gradle
@@ -203,6 +203,8 @@ dependencies {
 
     api "com.google.code.findbugs:annotations:${libraries.findbugs}"
 
+    api "com.squareup.okhttp3:mockwebserver:${libraries.okhttp}"
+
     // Java test dependencies
     testImplementation "junit:junit:${libraries.junit}"
     testImplementation "org.mockito:mockito-core:${libraries.mockito}"
@@ -215,7 +217,6 @@ dependencies {
     // Android test dependencies
     androidTestImplementation "commons-logging:commons-logging:1.2"
     androidTestImplementation "org.mockito:mockito-core:${libraries.mockito}"
-    androidTestImplementation "com.squareup.okhttp3:mockwebserver:${libraries.okhttp}"
     androidTestImplementation "android.arch.core:core-testing:${libraries.coretesting}"
     androidTestImplementation "com.jraska.livedata:testing:${libraries.livedatatesting}"
     androidTestImplementation "androidx.test:runner:${libraries.testrunner}"
