+++ core/src/androidTest/java/org/hisp/dhis/android/core/utils/integration/mock/BaseMockIntegrationTest.java
@@ -31,7 +31,7 @@ package org.hisp.dhis.android.core.utils.integration.mock;
 import org.hisp.dhis.android.core.D2;
 import org.hisp.dhis.android.core.arch.api.internal.ServerURLWrapper;
 import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter;
-import org.hisp.dhis.android.core.data.server.Dhis2MockServer;
+import org.hisp.dhis.android.core.mockwebserver.Dhis2MockServer;
 
 public abstract class BaseMockIntegrationTest {
